refactor(docs): rename geistdocs.tsx to site.config.tsx
Remove unused exports (suggestions, translations, basePath, siteId)
and update all imports to reference the new path.
