--- apps/docs/geistdocs.tsx
@@ -1,52 +1,0 @@
-export const Logo = () => (
-	<div>
-	  <p className="font-semibold text-xl tracking-tight">Shop</p>
-	</div>
-  );
-
-export const github = {
-	owner: "vercel",
-	repo: "shop",
-};
-
-export const nav = [
-	{
-		label: "Docs",
-		href: "/docs",
-	},
-	{
-		label: "Demo",
-		href: "https://vercel-shop.labs.vercel.dev",
-		target: "_blank",
-	},
-	{
-		label: "GitHub",
-		href: "https://github.com/vercel/shop",
-		target: "_blank",
-	},
-];
-
-export const suggestions = [
-	"What is Vercel Shop?",
-	"How can I localize my Vercel Shop site?",
-	"How do I deploy my Vercel Shop site?",
-];
-
-export const title = "Vercel Shop Documentation";
-
-export const prompt =
-	"You are a helpful assistant specializing in answering questions about Vercel Shop, the standard for Shopify development.";
-
-export const translations = {
-	en: {
-		displayName: "English",
-	},
-};
-
-export const basePath: string | undefined = undefined;
-
-/**
- * Unique identifier for this site, used in markdown request tracking analytics.
- * Each site using geistdocs should set this to a unique value (e.g. "ai-sdk-docs", "next-docs").
- */
-export const siteId: string | undefined = undefined;
